requirements.txt
cachetools==5.3.0
charset-normalizer==3.0.1
db-dtypes==1.0.5
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.2
google-cloud-bigquery-storage==2.18.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
install==1.3.5
mypy-extensions==0.4.3
numpy
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.0
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
